requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi>=0.100.0
pydantic>=2.0
fastapi-cli==0.0.5
Flask==3.0.3
google_search_results==2.4.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
outscraper==5.2.1
packaging==24.1
passlib==1.7.4
playwright==1.46.0
pluggy==1.5.0
# psycopg2-binary==2.9.9
pydantic_core==2.20.1
pyee==11.1.0
Pygments==2.18.0
pymongo==4.5.0
pytest==8.3.2
pytest-base-url==2.1.0
pytest-playwright==0.5.1
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.31.0
rich==13.8.0
serpapi==0.1.5
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.2
text-unidecode==1.3
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.23.2
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
Werkzeug==3.1.2